move inc/common/hal contents to inc/common/core
